+++ src/Attributes/Validation/MaxDigits.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Spatie\LaravelData\Attributes\Validation;
+
+use Attribute;
+use Spatie\LaravelData\Support\Validation\References\RouteParameterReference;
+
+#[Attribute(Attribute::TARGET_PROPERTY | Attribute::TARGET_PARAMETER)]
+class MaxDigits extends StringValidationAttribute
+{
+    public function __construct(protected int|RouteParameterReference $value)
+    {
+    }
+
+    public static function keyword(): string
+    {
+        return 'max_digits';
+    }
+
+    public function parameters(): array
+    {
+        return [$this->value];
+    }
+}