+++ src/Concerns/EnumerableMethods.php
@@ -23,7 +23,7 @@ trait EnumerableMethods
     {
         $cloned = clone $this;
 
-        $cloned->items = $cloned->items->map($through);
+        $cloned->items = $cloned->items->map(...func_get_args());
 
         return $cloned;
     }
@@ -35,7 +35,7 @@ trait EnumerableMethods
      */
     public function map(callable $map): static
     {
-        return $this->through($map);
+        return $this->through(...func_get_args());
     }
 
     /**
@@ -47,7 +47,21 @@ trait EnumerableMethods
     {
         $cloned = clone $this;
 
-        $cloned->items = $cloned->items->filter($filter);
+        $cloned->items = $cloned->items->filter(...func_get_args());
+
+        return $cloned;
+    }
+
+    /**
+     * @param callable(TValue): bool $filter
+     *
+     * @return static
+     */
+    public function reject(callable $filter): static
+    {
+        $cloned = clone $this;
+
+        $cloned->items = $cloned->items->reject(...func_get_args());
 
         return $cloned;
     }
@@ -62,7 +76,7 @@ trait EnumerableMethods
      */
     public function first(callable|null $callback = null, $default = null)
     {
-        return $this->items->first($callback, $default);
+        return $this->items->first(...func_get_args());
     }
 
     /**
@@ -75,7 +89,7 @@ trait EnumerableMethods
      */
     public function last(callable|null $callback = null, $default = null)
     {
-        return $this->items->last($callback, $default);
+        return $this->items->last(...func_get_args());
     }
 
     /**
@@ -85,7 +99,7 @@ trait EnumerableMethods
      */
     public function each(callable $callback): static
     {
-        $this->items->each($callback);
+        $this->items->each(...func_get_args());
 
         return $this;
     }
@@ -106,7 +120,7 @@ trait EnumerableMethods
     {
         $cloned = clone $this;
 
-        $cloned->items = $cloned->items->where($key, $operator, $value);
+        $cloned->items = $cloned->items->where(...func_get_args());
 
         return $cloned;
     }
@@ -122,7 +136,7 @@ trait EnumerableMethods
      */
     public function reduce(callable $callback, mixed $initial = null)
     {
-        return $this->items->reduce($callback, $initial);
+        return $this->items->reduce(...func_get_args());
     }
 
     /**